imports outside of src are not supported, moved globals back
